requirements.txt
apex
editdistance==0.6.2
einops==0.6.1
ipython==8.12.2
lightgbm==4.0.0
numpy
omegaconf==2.1.1
opencv_python==4.6.0.66
packaging==21.3
Pillow
PyYAML
regex==2021.11.10
scikit_learn==1.1.1
scipy==1.9.0
sentencepiece==0.1.99
six==1.16.0
tensorflow==2.13.0
torch==1.11.0
torchaudio==0.11.0
torchcrf==1.1.0
torchvision==0.12.0
tqdm==4.62.3
h5py
